requirements-dev.txt
asn1crypto==0.24.0
astroid==2.2.0
attrs==18.2.0
Automat==0.7.0
autopep8==1.4.3
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.6.1
cssselect==1.0.3
enum34==1.1.6
ffmpeg-python==0.2
fire==0.1.3
future==0.17.1
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
ipaddress==1.0.22
isort==4.3.9
lazy-object-proxy==1.3.1
lxml==4.3.1
mccabe==0.6.1
numpy==1.16.2
opencv-python==4.0.0.21
packaging==19.0
parsel==1.5.1
pep8==1.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.5.0
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pylint==2.3.0
pyOpenSSL==19.0.0
pyparsing==2.3.1
queuelib==1.5.0
requests==2.21.0
Scrapy==1.6.0
service-identity==18.1.0
six==1.12.0
tqdm==4.31.1
Twisted==19.2.1
typed-ast==1.3.1
urllib3==1.24.2
virtualenv==16.4.1
w3lib==1.20.0
wrapt==1.11.1
zope.interface==4.6.0
boto3==1.9.213